requirements.txt
alpaca-trade-api==3.1.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.3.1
matplotlib==3.8.4
requests==2.31.0
yfinance==0.2.40
schedule==1.2.1
